Updated with moving v fixed and source list
Included new code to mark each row and source as "moving" or "fixed depending on degree movement across first and last data-point. Nodding along slit not taken into account in this iteration, but the results seem to be good so far. Added new sheet creation named "Source List" that lists out a condensed version of the general log.