Adding the Score functions to the code
I added the math package becuase I needed a function out of it. I also added the code needed to get the scores for each target calibrator pair and pick the best one. I have made it compatible with the makelog function, because the function that is intended to be used takes the dictionary final and outputs a list of target calibrator pairs that are the best according to the algorithm used.